requirements.txt
fastapi==0.70.0
pydantic==1.8.2
uvicorn==0.15.0
pycryptodome==3.11.0
pytest
sqlalchemy ==1.4.27
alembic ==1.7.5
psycopg2-binary ==2.9.1